fix problems with maverick
