+++ test/e2e/app/config.go
@@ -20,6 +20,7 @@ type Config struct {
 	PrivValServer    string                      `toml:"privval_server"`
 	PrivValKey       string                      `toml:"privval_key"`
 	PrivValState     string                      `toml:"privval_state"`
+	Behavior         string                      `toml:"behaviors"`
 }
 
 // LoadConfig loads the configuration from disk.
