+++ test/e2e/docker/Dockerfile
@@ -18,7 +18,8 @@ RUN go mod download
 COPY . .
 RUN make build && cp build/tendermint /usr/bin/tendermint
 COPY test/e2e/docker/entrypoint* /usr/bin/
-RUN cd test/e2e && make app && cp build/app /usr/bin/app && make maverick && cp build/maverick /usr/bin/maverick
+RUN make maverick && cp build/maverick /usr/bin/maverick
+RUN cd test/e2e && make app && cp build/app /usr/bin/app
 
 # Set up runtime directory. We don't use a separate runtime image since we need
 # e.g. leveldb and rocksdb which are already installed in the build image.
