--- test/e2e/maverick/consensus/maverick_state.go
@@ -1,385 +1,0 @@
-package consensus
-
-import (
-	"fmt"
-	"reflect"
-	"sync"
-
-	cfg "github.com/tendermint/tendermint/config"
-	cstypes "github.com/tendermint/tendermint/consensus/types"
-	"github.com/tendermint/tendermint/crypto"
-	tmevents "github.com/tendermint/tendermint/libs/events"
-	"github.com/tendermint/tendermint/libs/service"
-	"github.com/tendermint/tendermint/p2p"
-	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
-	sm "github.com/tendermint/tendermint/state"
-	"github.com/tendermint/tendermint/types"
-)
-
-// State handles execution of the consensus algorithm.
-// It processes votes and proposals, and upon reaching agreement,
-// commits blocks to the chain and executes them against the application.
-// The internal state machine receives input from peers, the internal validator, and from a timer.
-type State struct {
-	service.BaseService
-
-	// config details
-	config        *cfg.ConsensusConfig
-	privValidator types.PrivValidator // for signing votes
-
-	// store blocks and commits
-	blockStore sm.BlockStore
-
-	// create and execute blocks
-	blockExec *sm.BlockExecutor
-
-	// notify us if txs are available
-	txNotifier txNotifier
-
-	// add evidence to the pool
-	// when it's detected
-	evpool evidencePool
-
-	// internal state
-	mtx sync.RWMutex
-	cstypes.RoundState
-	state sm.State // State until height-1.
-
-	// state changes may be triggered by: msgs from peers,
-	// msgs from ourself, or by timeouts
-	peerMsgQueue     chan msgInfo
-	internalMsgQueue chan msgInfo
-	timeoutTicker    TimeoutTicker
-	// privValidator pubkey, memoized for the duration of one block
-	// to avoid extra requests to HSM
-	privValidatorPubKey crypto.PubKey
-
-	// information about about added votes and block parts are written on this channel
-	// so statistics can be computed by reactor
-	statsMsgQueue chan msgInfo
-
-	// we use eventBus to trigger msg broadcasts in the reactor,
-	// and to notify external subscribers, eg. through a websocket
-	eventBus *types.EventBus
-
-	// a Write-Ahead Log ensures we can recover from any kind of crash
-	// and helps us avoid signing conflicting votes
-	wal          WAL
-	replayMode   bool // so we don't log signing errors during replay
-	doWALCatchup bool // determines if we even try to do the catchup
-
-	// for tests where we want to limit the number of transitions the state makes
-	nSteps int
-
-	// some functions can be overwritten for testing
-	decideProposal func(height int64, round int32)
-
-	// closed when we finish shutting down
-	done chan struct{}
-
-	// synchronous pubsub between consensus state and reactor.
-	// state only emits EventNewRoundStep and EventVote
-	evsw tmevents.EventSwitch
-
-	// for reporting metrics
-	metrics *Metrics
-
-	// behavior
-	behavior Behavior
-
-	// The height that the behaviour is executed
-	behaviorHeight int64
-
-	// the switch is passed to the state so that maveick behaviors can directly control which
-	// information they send to which nodes
-	sw *p2p.Switch
-}
-
-// StateOption sets an optional parameter on the State.
-type StateOption func(*State)
-
-// NewState returns a new State.
-func NewState(
-	config *cfg.ConsensusConfig,
-	state sm.State,
-	blockExec *sm.BlockExecutor,
-	blockStore sm.BlockStore,
-	txNotifier txNotifier,
-	evpool evidencePool,
-	behavior Behavior,
-	behaviorHeight int64,
-	options ...StateOption,
-) *State {
-	cs := &State{
-		config:               config,
-		blockExec:            blockExec,
-		blockStore:           blockStore,
-		txNotifier:           txNotifier,
-		peerMsgQueue:         make(chan msgInfo, msgQueueSize),
-		internalMsgQueue:     make(chan msgInfo, msgQueueSize),
-		timeoutTicker:        NewTimeoutTicker(),
-		statsMsgQueue:        make(chan msgInfo, msgQueueSize),
-		done:                 make(chan struct{}),
-		doWALCatchup:         true,
-		wal:                  nilWAL{},
-		evpool:               evpool,
-		evsw:                 tmevents.NewEventSwitch(),
-		metrics:              NopMetrics(),
-		behavior:            	behavior,
-		behaviorHeight: 			behaviorHeight,
-	}
-	// set function defaults (may be overwritten before calling Start)
-	cs.decideProposal = cs.defaultDecideProposal
-
-	// We have no votes, so reconstruct LastCommit from SeenCommit.
-	if state.LastBlockHeight > 0 {
-		cs.reconstructLastCommit(state)
-	}
-
-	cs.updateToState(state)
-
-	// Don't call scheduleRound0 yet.
-	// We do that upon Start().
-
-	cs.BaseService = *service.NewBaseService(nil, "State", cs)
-	for _, option := range options {
-		option(cs)
-	}
-	return cs
-}
-
-// state transitions on complete-proposal, 2/3-any, 2/3-one
-func (cs *State) handleMsg(mi msgInfo) {
-	cs.mtx.Lock()
-	defer cs.mtx.Unlock()
-
-	var (
-		added bool
-		err   error
-	)
-	msg, peerID := mi.Msg, mi.PeerID
-	switch msg := msg.(type) {
-	case *ProposalMessage:
-		// will not cause transition.
-		// once proposal is set, we can receive block parts
-		// err = cs.setProposal(msg.Proposal)
-		err = cs.behavior.ReceiveProposal(cs, msg.Proposal)
-	case *BlockPartMessage:
-		// if the proposal is complete, we'll enterPrevote or tryFinalizeCommit
-		added, err = cs.addProposalBlockPart(msg, peerID)
-		if added {
-			cs.statsMsgQueue <- mi
-		}
-
-		if err != nil && msg.Round != cs.Round {
-			cs.Logger.Debug(
-				"Received block part from wrong round",
-				"height",
-				cs.Height,
-				"csRound",
-				cs.Round,
-				"blockRound",
-				msg.Round)
-			err = nil
-		}
-	case *VoteMessage:
-		// attempt to add the vote and dupeout the validator if its a duplicate signature
-		// if the vote gives us a 2/3-any or 2/3-one, we transition
-		added, err = cs.tryAddVote(msg.Vote, peerID)
-		if added {
-			cs.statsMsgQueue <- mi
-		}
-
-		// if err == ErrAddingVote {
-		// TODO: punish peer
-		// We probably don't want to stop the peer here. The vote does not
-		// necessarily comes from a malicious peer but can be just broadcasted by
-		// a typical peer.
-		// https://github.com/tendermint/tendermint/issues/1281
-		// }
-
-		// NOTE: the vote is broadcast to peers by the reactor listening
-		// for vote events
-
-		// TODO: If rs.Height == vote.Height && rs.Round < vote.Round,
-		// the peer is sending us CatchupCommit precommits.
-		// We could make note of this and help filter in broadcastHasVoteMessage().
-	default:
-		cs.Logger.Error("Unknown msg type", "type", reflect.TypeOf(msg))
-		return
-	}
-
-	if err != nil {
-		cs.Logger.Error("Error with msg", "height", cs.Height, "round", cs.Round,
-			"peer", peerID, "err", err, "msg", msg)
-	}
-}
-
-// Enter (CreateEmptyBlocks): from enterNewRound(height,round)
-// Enter (CreateEmptyBlocks, CreateEmptyBlocksInterval > 0 ):
-// 		after enterNewRound(height,round), after timeout of CreateEmptyBlocksInterval
-// Enter (!CreateEmptyBlocks) : after enterNewRound(height,round), once txs are in the mempool
-func (cs *State) enterPropose(height int64, round int32) {
-	logger := cs.Logger.With("height", height, "round", round)
-
-	if cs.Height != height || round < cs.Round || (cs.Round == round && cstypes.RoundStepPropose <= cs.Step) {
-		logger.Debug(fmt.Sprintf(
-			"enterPropose(%v/%v): Invalid args. Current step: %v/%v/%v",
-			height,
-			round,
-			cs.Height,
-			cs.Round,
-			cs.Step))
-		return
-	}
-	logger.Info(fmt.Sprintf("enterPropose(%v/%v). Current: %v/%v/%v", height, round, cs.Height, cs.Round, cs.Step))
-
-	defer func() {
-		// Done enterPropose:
-		cs.updateRoundStep(round, cstypes.RoundStepPropose)
-		cs.newStep()
-
-		// If we have the whole proposal + POL, then goto Prevote now.
-		// else, we'll enterPrevote when the rest of the proposal is received (in AddProposalBlockPart),
-		// or else after timeoutPropose
-		if cs.isProposalComplete() {
-			cs.enterPrevote(height, cs.Round)
-		}
-	}()
-
-	cs.behavior.EnterPropose(cs, height, round)
-}
-
-// Enter: `timeoutPropose` after entering Propose.
-// Enter: proposal block and POL is ready.
-// Prevote for LockedBlock if we're locked, or ProposalBlock if valid.
-// Otherwise vote nil.
-func (cs *State) enterPrevote(height int64, round int32) {
-	if cs.Height != height || round < cs.Round || (cs.Round == round && cstypes.RoundStepPrevote <= cs.Step) {
-		cs.Logger.Debug(fmt.Sprintf(
-			"enterPrevote(%v/%v): Invalid args. Current step: %v/%v/%v",
-			height,
-			round,
-			cs.Height,
-			cs.Round,
-			cs.Step))
-		return
-	}
-
-	defer func() {
-		// Done enterPrevote:
-		cs.updateRoundStep(round, cstypes.RoundStepPrevote)
-		cs.newStep()
-	}()
-
-	cs.Logger.Info(fmt.Sprintf("enterPrevote(%v/%v). Current: %v/%v/%v", height, round, cs.Height, cs.Round, cs.Step))
-
-	// Sign and broadcast vote as necessary
-	cs.behavior.EnterPrevote(cs, height, round)
-
-	// Once `addVote` hits any +2/3 prevotes, we will go to PrevoteWait
-	// (so we have more time to try and collect +2/3 prevotes for a single block)
-}
-
-// Enter: `timeoutPrevote` after any +2/3 prevotes.
-// Enter: `timeoutPrecommit` after any +2/3 precommits.
-// Enter: +2/3 precomits for block or nil.
-// Lock & precommit the ProposalBlock if we have enough prevotes for it (a POL in this round)
-// else, unlock an existing lock and precommit nil if +2/3 of prevotes were nil,
-// else, precommit nil otherwise.
-func (cs *State) enterPrecommit(height int64, round int32) {
-	logger := cs.Logger.With("height", height, "round", round)
-
-	if cs.Height != height || round < cs.Round || (cs.Round == round && cstypes.RoundStepPrecommit <= cs.Step) {
-		logger.Debug(fmt.Sprintf(
-			"enterPrecommit(%v/%v): Invalid args. Current step: %v/%v/%v",
-			height,
-			round,
-			cs.Height,
-			cs.Round,
-			cs.Step))
-		return
-	}
-
-	logger.Info(fmt.Sprintf("enterPrecommit(%v/%v). Current: %v/%v/%v", height, round, cs.Height, cs.Round, cs.Step))
-
-	defer func() {
-		// Done enterPrecommit:
-		cs.updateRoundStep(round, cstypes.RoundStepPrecommit)
-		cs.newStep()
-	}()
-
-	cs.behavior.EnterPrecommit(cs, height, round)
-}
-
-func (cs *State) addVote(
-	vote *types.Vote,
-	peerID p2p.ID) (added bool, err error) {
-	cs.Logger.Debug(
-		"addVote",
-		"voteHeight",
-		vote.Height,
-		"voteType",
-		vote.Type,
-		"valIndex",
-		vote.ValidatorIndex,
-		"csHeight",
-		cs.Height,
-	)
-
-	// A precommit for the previous height?
-	// These come in while we wait timeoutCommit
-	if vote.Height+1 == cs.Height && vote.Type == tmproto.PrecommitType {
-		if cs.Step != cstypes.RoundStepNewHeight {
-			// Late precommit at prior height is ignored
-			cs.Logger.Debug("Precommit vote came in after commit timeout and has been ignored", "vote", vote)
-			return
-		}
-		added, err = cs.LastCommit.AddVote(vote)
-		if !added {
-			return
-		}
-
-		cs.Logger.Info(fmt.Sprintf("Added to lastPrecommits: %v", cs.LastCommit.StringShort()))
-		cs.eventBus.PublishEventVote(types.EventDataVote{Vote: vote})
-		cs.evsw.FireEvent(types.EventVote, vote)
-
-		// if we can skip timeoutCommit and have all the votes now,
-		if cs.config.SkipTimeoutCommit && cs.LastCommit.HasAll() {
-			// go straight to new round (skip timeout commit)
-			// cs.scheduleTimeout(time.Duration(0), cs.Height, 0, cstypes.RoundStepNewHeight)
-			cs.enterNewRound(cs.Height, 0)
-		}
-
-		return
-	}
-
-	// Height mismatch is ignored.
-	// Not necessarily a bad peer, but not favourable behaviour.
-	if vote.Height != cs.Height {
-		cs.Logger.Info("Vote ignored and not added", "voteHeight", vote.Height, "csHeight", cs.Height, "peerID", peerID)
-		return
-	}
-
-	added, err = cs.Votes.AddVote(vote, peerID)
-	if !added {
-		// Either duplicate, or error upon cs.Votes.AddByIndex()
-		return
-	}
-
-	cs.eventBus.PublishEventVote(types.EventDataVote{Vote: vote})
-	cs.evsw.FireEvent(types.EventVote, vote)
-
-	switch vote.Type {
-	case tmproto.PrevoteType:
-		cs.behavior.ReceivePrevote(cs, vote)
-
-	case tmproto.PrecommitType:
-		cs.behavior.ReceivePrecommit(cs, vote)
-
-	default:
-		panic(fmt.Sprintf("Unexpected vote type %v", vote.Type))
-	}
-
-	return added, err
-}
